Use the symeig package for normal mode calculations if it is installed
Use symeig (http://mdp-toolkit.sourceforge.net/symeig.html) if it is
installed and if NumPy is used (symeig works only with NumPy). Symeig
uses the LAPACK routine dsyevr, which uses less memory than dsyevd. It
is also said to be faster, but in my tests on the Mac (using Apple's
accelerated BLAS/LAPACK from the Accelerate framework) it turned out
to be slightly slower.