Modified contiguousObjectOffset for Collections
When a Collection was written to a file, contiguousObjectOffset tried
to make the whole Collection contiguous. In many cases, this is not
possible, and then bonds were often broken. In the current version,
only bonded units (molecules) are made contiguous, which is possible
in all but exceptional cases (large cycles).